third_party/blink/web_tests/editing/pasteboard/drop-text-events-sideeffect-crash-expected.txt

Ensure safety on side-effect on drop-initiated TextEvent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testTargetEditable.innerHTML is 'initialValue'
PASS testTargetIFrameDocument.body.innerHTML is 'initialBody'
PASS successfullyParsed is true

TEST COMPLETE
